Moves Python code to a separate module.
