+++ onMoped/run.py
@@ -1,0 +1,6 @@
+from nav import *
+from onMoped.constant_speed import constant_speed
+from extremely_simple_acc_test import acc_test
+from simple_acc import acc_run_test
+
+init()
