--- run.py
@@ -1,11 +1,0 @@
-import nav as n
-from nav import *
-from nav1 import whole4, pause, cont
-from driving import stop, drive, steer
-from constant_speed import constant_speed
-from acc import activate_acc, on, acc, acc_on
-from extremely_simple_acc_test import acc_test
-from simple_acc import acc_run_test
-from take_picture import test_steer
-
-init()
